+++ ProyectoAccesoriosNauticos/src/Modelo/Cliente.java
@@ -1,11 +1,5 @@
 package Modelo;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
 /**
  *
@@ -83,45 +77,5 @@ public class Cliente {
                 "\tTelefono: " + this.telefono; 
     }
     
-    public static List<Cliente> leer(){
-        // Variales a utilizar
-        List<Cliente> lista = new ArrayList<Cliente>();
-        String registro = "";
-        String[] atrregistros = null;
-        ArrayList<String> listclientes = null;
-        File fileclientes = null;
-        FileReader filer = null;
-        BufferedReader bufferr = null;
-        
-        // Manejo de archivos
-        try { 
-            // Archivos y buffer
-            fileclientes = new File(".\\src\\Proyecto\\Clientes.txt");
-            filer = new FileReader(fileclientes);
-            bufferr = new BufferedReader(filer);
-            
-            // Guardar todos los registros
-            while((registro=bufferr.readLine())!=null){
-                atrregistros = registro.split(",");
-                listclientes = new ArrayList<String>(Arrays.asList(atrregistros));
-                // Conertimos los datos y agregamos el objeto Producto
-                lista.add(new Cliente(
-                        Integer.parseInt(listclientes.get(0)),
-                        Integer.parseInt(listclientes.get(1)),
-                        listclientes.get(2),
-                        listclientes.get(3),
-                        listclientes.get(4)
-                ));
-            }
-            
-            // Cerrar el archivo
-            bufferr.close();
-            return lista;
-            
-        } catch (Exception e) {
-            System.out.println("Ha ocurido un error con el archivo: " + e);
-        }
-        return lista;
-    }   
-    
+     
 }
