+++ ProyectoAccesoriosNauticos/src/Modelo/Salida.java
@@ -1,15 +1,8 @@
 package Modelo;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Calendar;
+
 import java.util.Date;
-import java.util.List;
+
 
 /**
  *
@@ -46,76 +39,5 @@ public class Salida extends Pedido{
                 "\tMPagar: " + this.monto_pagar + "\tGanancia: " + this.ganancia; 
     }
     
-    // Leer datos de un archivo
-    public static List<Salida> leer(List<Producto> lista_productos){
-        // Variales a utilizar
-        List<Salida> lista = new ArrayList<Salida>();
-        DateFormat formateador= new SimpleDateFormat("dd/MM/yyyy");
-        String registro = "";
-        String[] atrregistros = null;
-        ArrayList<String> listsalida = null;
-        File filesalida = null;
-        FileReader filer = null;
-        BufferedReader bufferr = null;
-        BufferedReader buffersalida = null;
-        Producto prod = null;
-        
-        // Manejo de archivos
-        try { 
-            // Archivos y buffer
-            filesalida = new File(".\\src\\Proyecto\\Salidas.txt");
-            filer = new FileReader(filesalida);
-            bufferr = new BufferedReader(filer);
-            // Variables para fecha
-            Date fecha_e = new Date();
-            Date fecha_r = new Date();
-            // Guardar todos los registros
-            while((registro=bufferr.readLine())!=null){
-                atrregistros = registro.split(",");
-                listsalida = new ArrayList<String>(Arrays.asList(atrregistros));
-                // Verificamos si hay una fecha que ingresar
-                if (listsalida.get(1).equalsIgnoreCase("")) fecha_e = null;
-                else fecha_e = formateador.parse(listsalida.get(1));
-                if (listsalida.get(2).equalsIgnoreCase("")) fecha_r = null;
-                else fecha_r = formateador.parse(listsalida.get(2));
-                // Conertimos los datos y agregamos el objeto Productos
-                lista.add(new Salida(
-                        Integer.parseInt(listsalida.get(0)),
-                        fecha_e,
-                        fecha_r,
-                        Double.parseDouble(listsalida.get(3)),
-                        Double.parseDouble(listsalida.get(4))
-                ));
-            }
-            
-            // Cerrar el archivo
-            bufferr.close();
-            
-            // Recorremos la lista de entradas para ingresarle los productos
-            for (Salida salida: lista){
-                // Leer archivo de las entradas
-                filesalida = new File(".\\src\\Proyecto\\SalProd.txt");
-                filer = new FileReader(filesalida);
-                buffersalida = new BufferedReader(filer);   
-                // Guardar todos los registros
-                while((registro=buffersalida.readLine())!=null){
-                    atrregistros = registro.split(",");
-                    listsalida = new ArrayList<String>(Arrays.asList(atrregistros));
-                    // Verificamos si e producto es de 
-                    if (salida.getCod() == Integer.parseInt(listsalida.get(1))){
-                        // Bucamos el producto y lo agregamos
-                        prod = Producto.getProducto(lista_productos, Integer.parseInt(listsalida.get(0)));
-                        salida.setProductos(prod);
-                    }
-                }
-                
-                buffersalida.close();
-            }
-            return lista;
-            
-        } catch (Exception e) {
-            System.out.println("Ha ocurido un error con el archivo: " + e);
-        }
-        return lista;
-    }
+ 
 }
